feat(Editor): create useEditor hook
