+++ packages/hedgehog-lab/src/components/Snippet/Save/SaveState.tsx
@@ -2,7 +2,7 @@ import {Alert, IconButton} from "@mui/material";
 import {FiberManualRecord} from "@mui/icons-material";
 import React from "react";
 import {useRecoilValue} from "recoil";
-import {codeSavingFlagState} from "../../YourCode/RYourCodeStates";
+import {codeSavingFlagState} from "../../../states/RYourCodeStates";
 
 const SaveState = (): React.ReactElement => {
     const codeSavingFlag = useRecoilValue(codeSavingFlagState)
