+++ packages/hedgehog-lab/src/components/YourCode/Header/YourCodeHeader.tsx
@@ -2,17 +2,15 @@ import {Box, Button} from "@mui/material";
 import {PlayCircleOutline, StopCircleOutlined} from "@mui/icons-material";
 import {queryCache} from "react-query";
 import React, {useCallback} from "react";
-import {useRecoilValue} from "recoil";
-import {editorCodeState} from "../RYourCodeStates";
-import {COMPILE_AND_RUN_BUTTON_ID} from "../YourCode";
 import {compiler} from "../../../compiler";
 import SaveDialog from "../../Snippet/Save/SaveDialog";
 import {useParams} from "react-router-dom";
 import SaveState from "../../Snippet/Save/SaveState";
 import {useCompiler} from "../../../hooks/useCompilier";
+import {COMPILE_AND_RUN_BUTTON_ID, useEditor} from "../../../hooks/useEditor";
 
 const YourCodeHeader = (): React.ReactElement => {
-    const editorCode = useRecoilValue<string>(editorCodeState)
+    const [editorCode] = useEditor()
     const [setCompilerReFetch, isLoading] = useCompiler()
 
     const handleRunCode = useCallback(() => {
